requirements.txt
feedparser
requests
pyyaml
python-dotenv
google-genai
beautifulsoup4
fastapi
uvicorn
pydantic
duckduckgo-search
apscheduler
python-dateutil
